requirements.txt
# FaceTrack AI — Python dependencies
# Install with: pip install -r requirements.txt

# Web framework
flask>=3.0.0
flask-socketio>=5.3.6

# Environment variable loading
python-dotenv>=1.0.1

# Computer vision
opencv-contrib-python>=4.9.0

# Face detection (RetinaFace) + recognition (ArcFace)
insightface>=0.7.3
onnxruntime>=1.17.0

# Numerical / ML
numpy>=1.26.0
scikit-learn>=1.4.0
Pillow>=10.2.0

# Reporting
reportlab>=4.1.0
openpyxl>=3.1.2
pandas>=2.2.0

# Utilities
Werkzeug>=3.0.1
